Fixed if statement on line 501
`Line 501` had two independent conditions for `if isinstance(ld_coeffs, list)` or `if isinstance(ld_coeffs,np.ndarray)`.  If the input `ld_coeffs` was indeed a `list` or a `ndarray`, then it would trigger `True` and return `ld_coeff` as a `list(list)` or `list(ndarray)`.  

By changing `line 501` to `if not isinstance(ld_coeffs, (list, np.ndarray)):`, a `list` or `ndarray` properly triggers `False` when `list` or `ndarray` is given.